refactor(grpc): extract EPD encode into a first-class EncodeStage
