+++ model_gateway/src/routers/grpc/regular/stages/chat/preparation.rs
@@ -262,10 +262,10 @@ impl ChatPreparationStage {
             request.ignore_eos,
         );
 
-        let mut processed_messages = processed_messages;
-        processed_messages.multimodal_intermediate = multimodal_intermediate;
-
-        // Store results in context
+        // Store results in context. The multimodal intermediate is owned by
+        // ProcessingState (not nested in PreparationOutput) so EPD's EncodeStage
+        // can borrow it before request building serializes it a second time.
+        ctx.state.multimodal_intermediate = multimodal_intermediate;
         ctx.state.preparation = Some(PreparationOutput::Chat {
             token_ids,
             processed_messages,
