+++ model_gateway/src/routers/grpc/regular/stages/chat/request_building.rs
@@ -82,21 +82,15 @@ impl PipelineStage for ChatRequestBuildingStage {
         // Build chat request
         let request_id = format!("chatcmpl-{}", Uuid::now_v7());
 
-        // Assemble backend-specific multimodal data now that the backend is known.
-        // In EPD, request building also mints the encode->prefill rooms and
-        // carries the encode dispatch plan into the execution plan.
-        let mut encode_plan = None;
-        let multimodal_data = if let Some(intermediate) = processed_messages.multimodal_intermediate
-        {
-            let planned_encode =
-                helpers::plan_epd_encode(&intermediate, clients, ctx.state.workers.as_ref())
-                    .map_err(|e| {
-                        error!(function = "ChatRequestBuildingStage::execute", error = %e, "Failed to plan EPD encode");
-                        error::bad_request("multimodal_not_supported", format!("{e}"))
-                    })?;
-            let is_encode_routed = planned_encode.is_some();
-            encode_plan = planned_encode;
+        // EncodeStage (EPD) already planned the encode rendezvous and stashed it
+        // on `encode_outputs`. Its presence selects the pixel-drop assembly path.
+        let is_encode_routed = ctx.state.encode_outputs.is_some();
 
+        // Assemble backend-specific multimodal data now that the backend is known.
+        // The intermediate is owned by ProcessingState; take it here for the
+        // prefill serialization (EncodeStage already borrowed it for the encode
+        // payload). When encode-routed, drop the prefill pixels.
+        let multimodal_data = if let Some(intermediate) = ctx.state.multimodal_intermediate.take() {
             let assembled = if is_encode_routed {
                 assemble_multimodal_data_after_encode(
                     intermediate,
@@ -156,16 +150,12 @@ impl PipelineStage for ChatRequestBuildingStage {
             }
         }
 
-        // EPD: request building minted the per-item encode bootstrap info and
-        // planned the encode-worker dispatch. Inject the bootstrap info into the
-        // prefill request; the dispatch plan stays attached to ExecutionPlan.
-        let encode_dispatch = if let Some(plan) = encode_plan {
-            let (bootstrap_info, dispatch) = plan.into_parts();
-            proto_request.set_encode_bootstrap_info(bootstrap_info);
-            Some(dispatch)
-        } else {
-            None
-        };
+        // EPD: EncodeStage minted the per-item encode bootstrap info. Inject it
+        // into the prefill request; the dispatch plan stays on `encode_outputs`
+        // for request execution to take.
+        if let Some(outputs) = ctx.state.encode_outputs.as_ref() {
+            proto_request.set_encode_bootstrap_info(outputs.bootstrap_info.clone());
+        }
 
         // EPD: inject the prefill->decode KV rendezvous for backends that carry it
         // in the request. Runs before execute_parallel_pd clones the request, so
@@ -174,11 +164,7 @@ impl PipelineStage for ChatRequestBuildingStage {
             helpers::maybe_inject_pd_rendezvous(&mut proto_request, workers);
         }
 
-        ctx.state.execution_plan = Some(ExecutionPlan::generate(
-            self.plan_kind,
-            proto_request,
-            encode_dispatch,
-        ));
+        ctx.state.execution_plan = Some(ExecutionPlan::generate(self.plan_kind, proto_request));
         Ok(None)
     }
