+++ model_gateway/src/routers/grpc/regular/stages/messages/preparation.rs
@@ -296,10 +296,10 @@ impl MessagePreparationStage {
             false, // ignore_eos — not available in Messages API
         );
 
-        let mut processed_messages = processed_messages;
-        processed_messages.multimodal_intermediate = multimodal_intermediate;
-
-        // Store results in context
+        // Store results in context. The multimodal intermediate is owned by
+        // ProcessingState (not nested in PreparationOutput) so EPD's EncodeStage
+        // can borrow it before request building serializes it a second time.
+        ctx.state.multimodal_intermediate = multimodal_intermediate;
         ctx.state.preparation = Some(PreparationOutput::Messages {
             token_ids,
             processed_messages,
